+++ src/classes/router/rateLimit.ts
@@ -1,4 +1,4 @@
-import { as, time, randomNum } from "rjutils-collection"
+import { as, time, randomNum, zValidate } from "rjutils-collection"
 import { ExcludeFrom } from "../../types/internal"
 
 export default class RouteRateLimit<Excluded extends (keyof RouteRateLimit)[] = []> {
@@ -15,6 +15,22 @@ export default class RouteRateLimit<Excluded extends (keyof RouteRateLimit)[] = 
 	}
 
 	/**
+	 * Set the Ratelimit Identifier
+	 * 
+	 * This is useful for when you want to make two code seperate rate limit rules act like one, just assign them the
+	 * same Identifier and they are interlinked. be wary, this *might* cause issues if the rate limits arent the same
+	 * rules.
+	 * @default randomNum(1,10000000)
+	 * @warn ONLY USE IF YOU KNOW WHAT YOU ARE DOING
+	 * @since 8.7.2
+	*/ @zValidate([ (z) => z.number() ])
+	public identifier(identifier: number): ExcludeFrom<RouteRateLimit<[...Excluded, 'identifier']>, [...Excluded, 'identifier']> {
+		this.data.sortTo = identifier
+
+		return as<any>(this)
+	}
+
+	/**
 	 * Set the Penalty when hitting a rate limit in ms
 	 * 
 	 * When the User hits the endpoint(s) more than `<maxHits>` in `<timeWindow>ms`, the penalty will be applied to
@@ -22,10 +38,11 @@ export default class RouteRateLimit<Excluded extends (keyof RouteRateLimit)[] = 
 	 * for the endpoint(s). If the User hits the endpoint(s) less than `<maxHits>` in `<timeWindow>ms`, the penalty wont be applied
 	 * and if `<timeWindow>ms` has passed, the limits will reset without any penalty applying.
 	 * 
-	 You can always prevent a request / message from counting towards the ratelimit by calling `<HTTPRequest | WSMessage>.skipRateLimit()`
+	 * You can always prevent a request / message from counting towards the ratelimit by calling `<HTTPRequest | WSMessage>.skipRateLimit()`
 	 * @default time(10).s()
 	 * @since 8.6.0
-	*/ public penalty(ms: number): ExcludeFrom<RouteRateLimit<[...Excluded, 'penalty']>, [...Excluded, 'penalty']> {
+	*/ @zValidate([ (z) => z.number() ])
+	public penalty(ms: number): ExcludeFrom<RouteRateLimit<[...Excluded, 'penalty']>, [...Excluded, 'penalty']> {
 		this.data.penalty = ms
 
 		return as<any>(this)
@@ -42,7 +59,8 @@ export default class RouteRateLimit<Excluded extends (keyof RouteRateLimit)[] = 
 	 * You can always prevent a request / message from counting towards the ratelimit by calling `<HTTPRequest | WSMessage>.skipRateLimit()`
 	 * @default time(10).s()
 	 * @since 8.6.0
-	*/ public window(ms: number): ExcludeFrom<RouteRateLimit<[...Excluded, 'window']>, [...Excluded, 'window']> {
+	*/ @zValidate([ (z) => z.number() ])
+	public window(ms: number): ExcludeFrom<RouteRateLimit<[...Excluded, 'window']>, [...Excluded, 'window']> {
 		this.data.timeWindow = ms
 
 		return as<any>(this)
@@ -60,7 +78,8 @@ export default class RouteRateLimit<Excluded extends (keyof RouteRateLimit)[] = 
 	 * You can always prevent a request / message from counting towards the ratelimit by calling `<HTTPRequest | WSMessage>.skipRateLimit()`
 	 * @default Infinity
 	 * @since 8.6.0
-	*/ public hits(amount: number): ExcludeFrom<RouteRateLimit<[...Excluded, 'hits']>, [...Excluded, 'hits']> {
+	*/ @zValidate([ (z) => z.number() ])
+	public hits(amount: number): ExcludeFrom<RouteRateLimit<[...Excluded, 'hits']>, [...Excluded, 'hits']> {
 		this.data.maxHits = amount
 
 		return as<any>(this)