+++ src/misc/methodsEnum.ts
@@ -1,13 +1,11 @@
-enum Methods {
-  CONNECT = 'CONNECT',
-  TRACE = 'TRACE',
-  OPTIONS = 'OPTIONS',
-  DELETE = 'DELETE',
-  PATCH = 'PATCH',
-  POST = 'POST',
-  HEAD = 'HEAD',
-  PUT = 'PUT',
-  GET = 'GET',
-}
-
-export default Methods
+export default Object.freeze({
+  CONNECT: 'CONNECT',
+	TRACT: 'TRACE',
+	OPTIONS: 'OPTIONS',
+	DELETE: 'DELETE',
+	PATCH: 'PATCH',
+	HEAD: 'HEAD',
+	POST: 'POST',
+	PUT: 'PUT',
+	GET: 'GET'
+})